+++ packages/swarm-extension/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+
+### Fixed
+
+- Fixed `/swarm run` reusing a separately discovered auth storage instead of the active session model registry's auth storage ([#1472](https://github.com/can1357/oh-my-pi/issues/1472)).